refactor(main): split main.ts into window, updater, preferences, and ipc-handlers modules
